+++ BigDataSort/IO/IDataReader.cs
@@ -6,7 +6,7 @@ namespace BigDataSort
 	public interface IDataReader<T> : IDisposable
 	{
 		bool IsEnd { get; }
-		IEnumerable<T> ReadChunk(int size);
+		IEnumerable<T> ReadChunk(long size);
 		T Read();
 	}
 }
